+++ tests/test_nethsm_other.py
@@ -139,3 +139,8 @@ def test_state_provision_add_user_get_random_data(nethsm_no_provision: NetHSM) -
         assert random_data1 != random_data2
         assert random_data1 != random_data3
         assert random_data2 != random_data3
+
+
+def test_cluster_diagnostics(nethsm: NetHSM) -> None:
+    diagnostics = nethsm.get_cluster_diagnostics()
+    assert not diagnostics.state.running